requirements.txt
BeautifulSoup==3.2.1
Django==1.7.4
MySQL-python==1.2.5
djangorestframework==3.0.5
django-cors-headers==1.0.0
django-filter==0.9.2
django-mailgun==0.2.2
djoser==0.2.1
httplib2==0.9
twilio==3.7.1
twitter==1.16.0
unittest2==0.8.0
